refactor Role.java
