+++ src/main/java/ls/electric/demo/common/users/controller/UserController.java
@@ -2,15 +2,12 @@ package ls.electric.demo.common.users.controller;
 
 import ls.electric.demo.common.users.service.UserService;
 import ls.electric.demo.common.users.service.dto.UserResponse;
-import ls.electric.demo.config.security.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/user")
