+++ src/main/java/ls/electric/demo/common/users/service/UserService.java
@@ -4,13 +4,12 @@ import lombok.RequiredArgsConstructor;
 import ls.electric.demo.common.users.domain.User;
 import ls.electric.demo.common.users.repository.UserRepository;
 import ls.electric.demo.common.users.service.dto.UserResponse;
-import ls.electric.demo.config.security.Role;
+import ls.electric.demo.config.security.role.Role;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.util.Arrays;
-import java.util.List;
 
 @RequiredArgsConstructor
 @Service
