+++ src/main/java/ls/electric/demo/config/security/WebFluxSecurityConfig.java
@@ -2,6 +2,7 @@ package ls.electric.demo.config.security;
 
 import lombok.AllArgsConstructor;
 import ls.electric.demo.config.security.jwt.SecurityContextRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,12 +20,14 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.server.SecurityWebFilterChain;
 import reactor.core.publisher.Mono;
 
-@AllArgsConstructor
 @EnableWebFluxSecurity
 @EnableReactiveMethodSecurity
 public class WebFluxSecurityConfig {
 
+    @Autowired
     private AuthenticationManager authenticationManager;
+
+    @Autowired
     private SecurityContextRepository securityContextRepository;
 
     @Bean
@@ -37,33 +40,17 @@ public class WebFluxSecurityConfig {
                         Mono.fromRunnable(() -> swe.getResponse().setStatusCode(HttpStatus.FORBIDDEN))
                 ).and()
                 .csrf().disable()
-                .formLogin().disable()
-                .httpBasic().disable()
+                .formLogin(Customizer.withDefaults())
+                .httpBasic(Customizer.withDefaults())
                 .authenticationManager(authenticationManager)
                 .securityContextRepository(securityContextRepository)
                 .authorizeExchange()
                 .pathMatchers(HttpMethod.OPTIONS).permitAll()
                 .pathMatchers("/api/login").permitAll()
+                .pathMatchers("/swagger-ui").permitAll()
                 .anyExchange().authenticated()
                 .and().build();
     }
-
-//    @Bean
-//    public MapReactiveUserDetailsService userDetailsService() {
-//        UserDetails user = User
-//                .withUsername("user")
-//                .password(passwordEncoder().encode("test"))
-//                .roles("USER")
-//                .build();
-//
-//        UserDetails admin = User
-//                .withUsername("admin")
-//                .password(passwordEncoder().encode("test"))
-//                .roles("ADMIN")
-//                .build();
-//
-//        return new MapReactiveUserDetailsService(user,admin);
-//    }
 
     @Bean
     public PasswordEncoder passwordEncoder(){
