+++ src/main/java/ls/electric/demo/config/security/role/Role.java
@@ -1,4 +1,4 @@
-package ls.electric.demo.config.security;
+package ls.electric.demo.config.security.role;
 
 public enum Role {
     USER, ADMIN, ANONYMOUS
